Clean up LC_CSR_Graph
Remove degree and hyper graph fields, which are specific to graph
specializations.

Delete dead code.
