+++ libcusp/include/galois/graphs/NewGeneric.h
@@ -306,8 +306,6 @@ public:
     fillMirrors();
     TfillMirrors.stop();
 
-    base_DistGraph::printStatistics();
-
     if (_edgeStateRounds > 1) {
       // reset edge load since we need exact same answers again
       resetEdgeLoad();
@@ -398,11 +396,6 @@ private:
       assignedThreadRanges[i] += startNode;
     }
 
-    // galois::gPrint("[", base_DistGraph::id, "] num local is ", numLocalNodes,
-    // "\n"); for (uint32_t i : assignedThreadRanges) {
-    //  galois::gPrint("[", base_DistGraph::id, "]", i , "\n");
-    //}
-
     auto toReturn = galois::runtime::makeSpecificRange(
         boost::counting_iterator<size_t>(startNode),
         boost::counting_iterator<size_t>(startNode + numLocalNodes),
@@ -436,14 +429,6 @@ private:
 
     galois::runtime::SpecificRange<boost::counting_iterator<size_t>> work =
         getSpecificThreadRange(bufGraph, rangeVector, start, end);
-
-    // galois::on_each([&] (unsigned i, unsigned j) {
-    //  galois::gPrint("[", base_DistGraph::id, " ", i, "] local range ",
-    //  *work.local_begin(), " ", *work.local_end(), "\n");
-    //});
-    // galois::PerThreadTimer<CUSP_PT_TIMER> ptt(
-    //  GRNAME, "Phase0DetNeighLocation_" + std::string(base_DistGraph::id)
-    //);
 
     // Step 2: loop over all local nodes, determine neighbor locations
     galois::do_all(
@@ -521,8 +506,8 @@ private:
                    (ghosts.size() - numLocal) / 64, " vs. total vector size ",
                    numToReserve / 2);
 
-    ghosts.resize(
-        0); // TODO: should not be used after this - refactor to make this clean
+    // TODO: should not be used after this - refactor to make this clean
+    ghosts.resize(0);
 
     mapSetupTimer.stop();
 
@@ -567,10 +552,6 @@ private:
       uint32_t sendingHost = p->first;
       // deserialize into neighbor bitsets
       galois::runtime::gDeserialize(p->second, syncNodes[sendingHost]);
-
-      // for (uint32_t i : ghosts[sendingHost].getOffsets()) {
-      //  galois::gDebug("[", base_DistGraph::id, "] ", i, " is set");
-      //}
     }
 
     p0BitsetCommTimer.stop();
@@ -848,12 +829,6 @@ private:
       std::vector<uint32_t> mastersToSend =
           getDataFromOffsets(offsetVector, dataVector);
 
-      // for (unsigned i : mastersToSend) {
-      //  galois::gDebug("[", base_DistGraph::id, "] gid ",
-      //                 i + base_DistGraph::gid2host[net.ID].first,
-      //                 " master send ", i);
-      //}
-      // assert it's a positive number
       assert(mastersToSend.size());
 
       size_t num_selected = toSync.count();
@@ -2995,8 +2970,7 @@ private:
       uint64_t gdst = gdst_vec[i++];
       uint32_t ldst = this->G2L(gdst);
       graph.constructEdge(cur++, ldst, gdata);
-      // TODO
-      // if ldst is an outgoing mirror, this is vertex cut
+      // TODO if ldst is an outgoing mirror, this is vertex cut
     }
   }
 
@@ -3011,46 +2985,9 @@ private:
       uint64_t gdst = gdst_vec[i++];
       uint32_t ldst = this->G2L(gdst);
       graph.constructEdge(cur++, ldst);
-      // TODO
-      // if ldst is an outgoing mirror, this is vertex cut
+      // TODO if ldst is an outgoing mirror, this is vertex cut
     }
   }
-
-  // public:
-  // virtual void boostSerializeLocalGraph(boost::archive::binary_oarchive& ar,
-  //                                      const unsigned int version = 0) const
-  //                                      {
-  //  // unsigned ints
-  //  ar << base_DistGraph::numNodes;
-
-  //  // partition specific
-  //  graphPartitioner->serializePartition(ar);
-
-  //  // maps and vectors
-  //  ar << base_DistGraph::localToGlobalVector;
-  //  ar << base_DistGraph::globalToLocalMap;
-  //}
-
-  // virtual void boostDeSerializeLocalGraph(boost::archive::binary_iarchive&
-  // ar,
-  //                                        const unsigned int version = 0) {
-  //  graphPartitioner = new Partitioner(base_DistGraph::id,
-  //  base_DistGraph::numHosts,
-  //                                     base_DistGraph::numGlobalNodes,
-  //                                     base_DistGraph::numGlobalEdges);
-
-  //  graphPartitioner->saveGIDToHost(base_DistGraph::gid2host);
-
-  //  // unsigned ints
-  //  ar >> base_DistGraph::numNodes;
-
-  //  // partition specific
-  //  graphPartitioner->deserializePartition(ar);
-
-  //  // maps and vectors
-  //  ar >> base_DistGraph::localToGlobalVector;
-  //  ar >> base_DistGraph::globalToLocalMap;
-  //}
 };
 
 // make GRNAME visible to public
