+++ libgalois/include/galois/graphs/LC_CSR_Hypergraph.h
@@ -517,14 +517,6 @@ public:
         galois::no_stats(), galois::steal());
   }
 
-  template <typename F>
-  ptrdiff_t partition_neighbors(GraphNode N, const F& func) {
-    auto beg = &edgeDst[*raw_begin(N)];
-    auto end = &edgeDst[*raw_end(N)];
-    auto mid = std::partition(beg, end, func);
-    return (mid - beg);
-  }
-
   void allocateFrom(FileGraph& graph) {
     numNodes = graph.size();
     numEdges = graph.sizeEdges();
