+++ libgalois/include/galois/graphs/LC_InOut_Graph.h
@@ -201,16 +201,6 @@ public:
     }
   }
 
-  template <typename F>
-  ptrdiff_t partition_in_neighbors(GraphNode N, const F& func) {
-
-    if (!asymmetric) {
-      return Super::partition_neighbors(N, func);
-    } else {
-      return inGraph.partition_neighbors(N, func);
-    }
-  }
-
   internal::InEdgesIterator<LC_InOut_Graph>
   in_edges(GraphNode N, MethodFlag mflag = MethodFlag::WRITE) {
     return internal::InEdgesIterator<LC_InOut_Graph>(*this, N, mflag);
