+++ tools/dist-graph-convert/dist-graph-convert-helpers.cpp
@@ -333,124 +333,6 @@ getChunkToHostMapping(const std::vector<uint64_t>& chunkCountsPrefixSum,
   return finalMapping;
 }
 
-std::vector<Uint64Pair>
-getChunkToHostMappingLinear(const std::vector<uint64_t>& chunkCountsPrefixSum,
-                            const std::vector<Uint64Pair>& chunkToNode) {
-  GALOIS_DIE("Currently does not work."); // TODO fix this function
-  uint64_t hostID        = galois::runtime::getSystemNetworkInterface().ID;
-  uint64_t totalNumHosts = galois::runtime::getSystemNetworkInterface().Num;
-
-  uint64_t totalWeight = chunkCountsPrefixSum.back();
-  uint64_t weightPerPartition =
-      (totalWeight + totalNumHosts - 1) / totalNumHosts;
-  uint64_t totalChunks = chunkCountsPrefixSum.size();
-
-  // TODO corner case handling (1 host, more hosts than chunks)
-
-  uint32_t currentHost  = 0;
-  uint64_t currentChunk = 0;
-
-  uint64_t accountedEdges = 0;
-
-  std::vector<uint64_t> hostRanges(totalNumHosts + 1);
-  hostRanges[0] = 0;
-
-  while (currentChunk < totalChunks && currentHost < totalNumHosts) {
-    uint32_t hostsRemaining = totalNumHosts - currentHost;
-    GALOIS_ASSERT(totalChunks - currentChunk >= hostsRemaining);
-
-    // Handle case where only 1 host left or 1-1 host-chunk mapping
-    if (hostsRemaining == 1) {
-      // assign the rest of chunks to last host
-      hostRanges[currentHost + 1] = totalChunks;
-    } else if ((totalChunks - currentChunk) == hostsRemaining) {
-      // one chunk to each host
-      for (unsigned i = 0; i < hostsRemaining; i++) {
-        hostRanges[++currentHost] = (++currentChunk);
-      }
-    }
-
-    // Number of edges in current chunk
-    uint64_t chunkEdges;
-    if (currentChunk > 0) {
-      chunkEdges = chunkCountsPrefixSum[currentChunk] -
-                   chunkCountsPrefixSum[currentChunk - 1];
-    } else { // currentChunk == 0
-      chunkEdges = chunkCountsPrefixSum[0];
-    }
-
-    // Num edges division currently has not accounting chunkEdges into it
-    uint64_t edgeCountWithoutCurrent;
-    if (currentChunk > 0) {
-      edgeCountWithoutCurrent =
-          chunkCountsPrefixSum[currentChunk] - accountedEdges - chunkEdges;
-    } else { // currentChunk == 0
-      edgeCountWithoutCurrent = 0;
-    }
-
-    // If chunk edges is large, then don't add to current host unless host
-    // doesn't have much to begin with
-    if (chunkEdges > (3 * weightPerPartition / 4)) {
-      // If adding this chunk to current host too much, add to next host
-      // instead
-      if (edgeCountWithoutCurrent > (weightPerPartition / 4)) {
-        GALOIS_ASSERT(currentChunk != 0);
-
-        // assign to next host
-        // Beginning of next is current chunk
-        hostRanges[currentHost + 1] = currentChunk;
-        accountedEdges              = chunkCountsPrefixSum[currentChunk - 1];
-        currentHost++;
-        continue;
-      }
-    }
-
-    // otherwise handle regularly
-    uint64_t currentEdgeCount = edgeCountWithoutCurrent + chunkEdges;
-
-    if (currentEdgeCount >= weightPerPartition) {
-      hostRanges[++currentHost] = currentChunk + 1;
-      accountedEdges            = chunkCountsPrefixSum[currentChunk];
-    }
-
-    currentChunk++;
-  }
-
-  galois::gPrint("[", hostID, "] Done here\n");
-
-  GALOIS_ASSERT(hostRanges[0] == 0);
-  GALOIS_ASSERT(hostRanges[totalNumHosts] == totalChunks);
-
-  // handle pair creation
-  std::vector<Uint64Pair> finalMapping;
-  for (uint64_t h = 0; h < totalNumHosts; h++) {
-    uint64_t lowerChunk = hostRanges[h];
-    uint64_t upperChunk = hostRanges[h + 1];
-
-    uint64_t lowerNode = chunkToNode[lowerChunk].first;
-    uint64_t upperNode = chunkToNode[upperChunk].first;
-
-    if (hostID == 0) {
-      uint64_t edgeCount;
-      if (lowerChunk == upperChunk) {
-        edgeCount = 0;
-      } else if (lowerChunk == 0) {
-        edgeCount = chunkCountsPrefixSum[upperChunk - 1];
-      } else {
-        edgeCount = chunkCountsPrefixSum[upperChunk - 1] -
-                    chunkCountsPrefixSum[lowerChunk - 1];
-      }
-      std::cout << "Host " << h << " gets nodes " << lowerNode << " to "
-                << upperNode << " (count " << (upperNode - lowerNode)
-                << "), with " << edgeCount << " edges\n";
-    }
-
-    finalMapping.emplace_back(Uint64Pair(lowerNode, upperNode));
-  }
-
-  return finalMapping;
-}
-
 DoubleUint64Pair getNodesToReadFromGr(const std::string& inputGr) {
   uint32_t hostID        = galois::runtime::getSystemNetworkInterface().ID;
   uint32_t totalNumHosts = galois::runtime::getSystemNetworkInterface().Num;
